Fix the exponential output, name collisions and array subscripts
Sweeping and measuring rather than guessing, mostly in the AST writer.

The decompiler could not handle a shader that builds each value on the one
before. Ten instructions of that came to sixty thousand characters and
sixteen to forty-four million, because an expression read in two places is
written out at both. Naming one costs a line, and naming it also hides it
from the grouping that turns four dot products back into a matrix
multiply, so it is only done when the expression written out in full
crosses a budget - which no shader here does, so nothing moved.

That bounded the output, not the work: sixteen instructions still took 21
seconds and twenty never finished. Calls to ReduceDepthFirst tripled per
instruction, so what a node reduces to is remembered; then the counts were
flat and the cost was inside a call, and NodeVisitor turned out to be
walking once per path rather than once per node - 25 million visits at
depth 16. A hundred and fifty-two instructions now take under a second.

Names that shadow each other: the output variable is called o and so was a
constant buffer; the input struct is called i and so is the first loop
counter, so i.texcoord read the counter. Both latent, neither reachable by
any fixture.

Struct members were named by component rather than by name, a scalar
member took a swizzle, and constant arrays were named without the element
- all four registers of float4 m[4] read as m. A matrix array needs both
subscripts, since its register offset counts rows: transpose(bones[a0/4])[1].

GetParamInt returned float, so an integer immediate was written as
-2,1474836E+09 - in the machine's own decimal separator, into a fixture.
udiv writes two registers and HasDestination cannot say so, so neither was
declared.

component_chain is fixed and no longer a known failure. Two bugs:
UseLoopVariablesInBody set an assignment's variable without rewriting what
already read the old one, and grouping ran before ordering, so two
assignments needing a third between them were merged.

Three tests added. NodeGraphConsistencyTests checks every edge is recorded
both ways, after finding a bypassed phi keeping its consumers.
RoundTripCostTests measures what a round trip costs, which is the only
thing that sees a defect that compiles - I wrote eleven reasons into it
from the numbers and most were wrong, so each entry now says whether it is
the decompiler or fxc. TempAssignmentOrderTests covers an ordering that
was not one: a partial order handed to a comparison sort put a variable's
use before its assignment, on four elements but not forty.

Known instruction failures are down from nine to two, and both of those
are recorded with why the obvious fix is worse than the bug.

+++ DirectXShaderModel/D3D10Instruction.cs
@@ -521,12 +521,12 @@ public class D3D10Instruction : Instruction
         return BitConverter.ToSingle(GetOperandValueBytes(index, componentIndex), 0);
     }
 
-    public override float GetParamInt(int index)
+    public override int GetParamInt(int index)
     {
         return BitConverter.ToInt32(GetOperandValueBytes(index, 0), 0);
     }
 
-    public float GetParamInt(int index, int componentIndex)
+    public int GetParamInt(int index, int componentIndex)
     {
         return BitConverter.ToInt32(GetOperandValueBytes(index, componentIndex), 0);
     }
